Added New Years Chaos
This is unfinished and also in Javascript. It runs for some test cases sometimes kind of... I also need to refine the algorithm a bit. This is a lot of fun. Oh, and I want to write the soution in Python. I had a lot of help from a friend in writing this in JS.